+++ sshd-common/src/main/java/org/apache/sshd/common/config/keys/impl/AbstractKeyEntryResolver.java
@@ -54,6 +54,6 @@ public abstract class AbstractKeyEntryResolver<PUB extends PublicKey, PRV extend
 
     @Override
     public String toString() {
-        return getPublicKeyType().getSimpleName() + ": " + getSupportedTypeNames();
+        return getPublicKeyType().getSimpleName() + ": " + getSupportedKeyTypes();
     }
 }